+++ CHANGELOG.md
@@ -5,7 +5,9 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how
 to structure this file.  This project does **not** adhere to [Semantic
 Versioning](http://semver.org/)!
 
-## [Unreleased]
+## [Unreleased]s
+
+### 0.2.0 – 2017-06-12
 ### Added
 - Add `Vale: Lint workspace` command to run vale on all files in the current
   workspace, see <https://github.com/lunaryorn/vscode-vale/issues/1>.
@@ -20,7 +22,7 @@ Versioning](http://semver.org/)!
 
 ## 0.1.1 – 2017-06-08
 ### Fixed
-- Use the right marketplace description.
+- Use a proper marketplace description.
 
 ## 0.1.0 – 2017-06-08
 This is the first release.